chore: testing other implementations and minor improvements
Implemented a script to save preview of assset as PNG (for icon). Run
multiple tests to identify why trigger press was being registered as
button A press. Found the problem in the CameraRig or its descendants.
Solved in new scene InventoryVRTest, but lost interactivity with
objects. Some additional clean up was made.